# openapi.yaml
openapi: 3.0.3
info:
  title: Delivery Service API
  version: 1.0.0
  description: |
    OpenAPI specification generated from existing Spring Boot controllers and DTOs.
    Includes only routes found in code under `src/main/kotlin`.
...
paths:
...
  /api/v1/admin/catalog/modifier-groups:
    get:
      tags: [Catalog Admin]
      operationId: getAdminModifierGroups
      summary: List modifier groups
      security:
        - bearerAuth: []
      parameters:
        - name: isActive
          in: query
          required: false
          description: When provided, filters modifier groups by activity flag.
          schema:
            type: boolean
      responses:
        '200':
          description: Modifier groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModifierGroupResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Catalog Admin]
      operationId: upsertModifierGroup
      summary: Create or update modifier group
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertModifierGroupRequest'
      responses:
        '200':
          description: Saved modifier group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModifierGroupResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "JWT access token in `Authorization: Bearer <token>`"
    installIdHeader:
      type: apiKey
      in: header
      name: X-Install-Id
      description: Guest actor identifier used by `CurrentActorArgumentResolver`
...
  responses:
    BadRequestError:
      description: Request validation failed or business rule violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    UnauthorizedError:
      description: Authentication required or invalid token/header
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    ForbiddenError:
      description: Authenticated but not allowed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
  schemas:
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum: [VALIDATION_ERROR, UNAUTHORIZED, FORBIDDEN, NOT_FOUND, INTERNAL_ERROR]
        message:
          type: string
        traceId:
          type: string
          nullable: true
          description: Request trace id (`X-Trace-Id`) when available.
        details:
          type: object
          nullable: true
          additionalProperties: true
          # TODO: concrete map shape varies by exception type; only stable contract is object map.
...
    ModifierGroupResponse:
      type: object
      required: [id, code, name, minSelected, maxSelected, isRequired, isActive, sortOrder, options]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        minSelected:
          type: integer
          format: int32
        maxSelected:
          type: integer
          format: int32
        isRequired:
          type: boolean
        isActive:
          type: boolean
        sortOrder:
          type: integer
          format: int32
        options:
          type: array
          items:
            $ref: '#/components/schemas/ModifierOptionResponse'
    ModifierOptionResponse:
      type: object
      required: [id, code, name, priceType, price, applicationScope, isDefault, isActive, sortOrder]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        priceType:
          $ref: '#/components/schemas/ModifierPriceType'
        price:
          type: integer
          format: int64
        applicationScope:
          $ref: '#/components/schemas/ModifierApplicationScope'
        isDefault:
          type: boolean
        isActive:
          type: boolean
        sortOrder:
          type: integer
          format: int32
...
    UpsertModifierGroupRequest:
      type: object
      required: [code, name, minSelected, maxSelected]
      properties:
        id:
          type: string
          format: uuid
          nullable: true
        code:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        minSelected:
          type: integer
          format: int32
          minimum: 0
        maxSelected:
          type: integer
          format: int32
          minimum: 1
        isRequired:
          type: boolean
          default: false
        isActive:
          type: boolean
          default: true
        sortOrder:
          type: integer
          format: int32
          default: 0
        options:
          type: array
          items:
            $ref: '#/components/schemas/UpsertModifierOptionRequest'
    UpsertModifierOptionRequest:
      type: object
      required: [code, name, priceType, applicationScope]
      properties:
        code:
          type: string
          minLength: 1
        name:
          type: string
          minLength: 1
        description:
          type: string
          nullable: true
        priceType:
          $ref: '#/components/schemas/ModifierPriceType'
        price:
          type: integer
          format: int64
          minimum: 0
        applicationScope:
          $ref: '#/components/schemas/ModifierApplicationScope'
        isDefault:
          type: boolean
          default: false
        isActive:
          type: boolean
          default: true
        sortOrder:
          type: integer
          format: int32
          default: 0
...
    ModifierPriceType:
      type: string
      enum: [FIXED, FREE]
    ModifierApplicationScope:
      type: string
      enum: [PER_ITEM, PER_LINE]
...